Add prompts package, enhanced judges, sensory forecasts, and state metadata
Consolidate all prompt templates into src/minsky/prompts/ package.
Add architecture preamble and message history context to judge evaluations.
Judge all three rooms (including Sensory). Add self-supervised sensory
forecast system that generates training pairs without judges. Switch
summarizer from RWKV to DeepSeek judge model API. Use format_t5_prompt()
everywhere. Save global_step in RWKV state metadata for resume across runs.

+++ scripts/download_model.py
@@ -118,11 +118,11 @@ def download_t5() -> Path:
 
     MODELS_DIR.mkdir(parents=True, exist_ok=True)
 
-    # Download and save to local path
+    # Download and save to local path (float32 for T5)
     processor = AutoProcessor.from_pretrained(T5_MODEL)
     model = AutoModelForSeq2SeqLM.from_pretrained(
         T5_MODEL,
-        torch_dtype=torch.bfloat16,
+        torch_dtype=torch.float32,
     )
 
     # Save locally
